Moved topics! macro to dedicated module
